Wires in multiplayer! (nonfunctional wires)
position of gameobjects stored as doubles instead of int now, position of blocks are int
Grid class added
mouse inputs fixed
BlockType class added
Scene concurrent modification fixed
Block class added
BlockPosition class added
Cursor class added
Player can left click now and place wires
isPressed changed to isKeyPressed
Wire class added
New packet called BlockActionPacket added
BlockActionPacket class added
MovePlayerPacket and NewPlayerPacket have double type for x,y instead of int type
Server got all the same updates network wise
ServerGrid class now added. The server stores block positions now.
Only action sent right now is the placing a block action, removing a block isn't implemented yet.

+++ GKSim/src/core/GameObject.java
@@ -1,16 +1,17 @@
 package core;
 
+import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.util.UUID;
 
 public abstract class GameObject
 {
     protected Game game;
-    protected int x, y;
+    protected double x, y;
     protected BufferedImage image;
     protected UUID uuid;
 
-    public GameObject(Game game, UUID uuid, int x, int y)
+    public GameObject(Game game, UUID uuid, double x, double y)
     {
         this.game = game;
         this.x = x;
@@ -39,22 +40,22 @@ public abstract class GameObject
         return uuid;
     }
 
-    public void setX(int x)
+    public void setX(double x)
     {
         this.x = x;
     }
 
-    public int getX()
+    public double getX()
     {
         return x;
     }
 
-    public void setY(int y)
+    public void setY(double y)
     {
         this.y = y;
     }
 
-    public int getY()
+    public double getY()
     {
         return y;
     }
